requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery[redis]
redis==5.0.1
celery==5.3.6
mysqlclient==2.0.3
